+++ lib/ui/events/CompositeEventsDetailPanel.dart
@@ -20,6 +20,7 @@ import 'package:flutter/material.dart';
 import 'package:geolocator/geolocator.dart' as Core;
 import 'package:illinois/ext/Event.dart';
 import 'package:illinois/ext/Explore.dart';
+import 'package:illinois/ui/widgets/SmallRoundedButton.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/model/RecentItem.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
@@ -32,7 +33,7 @@ import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:illinois/ui/events/EventsSchedulePanel.dart';
 import 'package:illinois/ui/explore/ExploreEventDetailPanel.dart';
 import 'package:illinois/ui/widgets/PrivacyTicketsDialog.dart';
-import 'package:illinois/ui/widgets/ScalableWidgets.dart';
+import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:illinois/ui/widgets/SectionTitlePrimary.dart';
 
 import 'package:illinois/service/RecentItems.dart';
@@ -43,7 +44,6 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/explore/ExploreConvergeDetailItem.dart';
 import 'package:illinois/ui/widgets/TabBarWidget.dart';
-import 'package:illinois/ui/widgets/RoundedButton.dart';
 
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:sprintf/sprintf.dart';
@@ -464,14 +464,13 @@ class _CompositeEventsDetailPanelState extends State<CompositeEventsDetailPanel>
     String? ticketsUrl = widget.parentEvent?.registrationUrl;
     bool getTicketsVisible = StringUtils.isNotEmpty(ticketsUrl);
 
-    String? websiteLabel = Localization().getStringEx('panel.explore_detail.button.visit_website.title', 'Visit website');
+    String websiteLabel = Localization().getStringEx('panel.explore_detail.button.visit_website.title', 'Visit website')!;
     String? websiteHint = Localization().getStringEx('panel.explore_detail.button.visit_website.hint', '');
 
     Widget visitWebsiteButton = (widget.parentEvent?.isSuperEvent ?? false) ?
     Visibility(visible: visitWebsiteVisible, child: SmallRoundedButton(
       label: websiteLabel,
       hint: websiteHint,
-      showChevron: true,
       borderColor: Styles().colors!.fillColorPrimary,
       onTap: () => _onTapVisitWebsite(titleUrl),),) :
     Visibility(visible: visitWebsiteVisible, child: RoundedButton(
@@ -490,7 +489,7 @@ class _CompositeEventsDetailPanelState extends State<CompositeEventsDetailPanel>
         visitWebsiteButton,
         Visibility(visible: visitWebsiteVisible, child: buttonsDivider),
         Visibility(visible: getTicketsVisible, child: RoundedButton(
-          label: Localization().getStringEx('panel.explore_detail.button.get_tickets.title', 'Get tickets'),
+          label: Localization().getStringEx('panel.explore_detail.button.get_tickets.title', 'Get tickets')!,
           hint: Localization().getStringEx('panel.explore_detail.button.get_tickets.hint', ''),
           backgroundColor: Colors.white,
           borderColor: Styles().colors!.fillColorSecondary,
@@ -570,28 +569,15 @@ class _CompositeEventsDetailPanelState extends State<CompositeEventsDetailPanel>
     Container(
         padding: EdgeInsets.symmetric(vertical: 10),
         child:
-        Stack(
-          children: [
-            ScalableRoundedButton(
-              label: Localization().getStringEx('panel.explore_detail.button.add_to_group.title', 'Add Event To Group') ,
-              hint: Localization().getStringEx('panel.explore_detail.button.add_to_group.hint', '') ,
-              backgroundColor: Colors.white,
-              borderColor: Styles().colors!.fillColorPrimary,
-              textColor: Styles().colors!.fillColorPrimary,
-              onTap: _onTapAddToGroup,
-            ),
-            Visibility(visible: _addToGroupInProgress,
-              child: Container(
-                height: 48,
-                child: Align(alignment: Alignment.center,
-                  child: SizedBox(height: 24, width: 24,
-                      child: CircularProgressIndicator(strokeWidth: 2, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors!.fillColorPrimary), )
-                  ),
-                ),
-              ),
-            ),
-          ],
-        )
+          RoundedButton(
+            label: Localization().getStringEx('panel.explore_detail.button.add_to_group.title', 'Add Event To Group')!,
+            hint: Localization().getStringEx('panel.explore_detail.button.add_to_group.hint', '') ,
+            backgroundColor: Colors.white,
+            borderColor: Styles().colors!.fillColorPrimary,
+            textColor: Styles().colors!.fillColorPrimary,
+            progress: _addToGroupInProgress,
+            onTap: _onTapAddToGroup,
+          ),
     );
   }