+++ lib/ui/groups/GroupPostCreatePanel.dart
@@ -4,7 +4,7 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
-import 'package:illinois/ui/widgets/RoundedButton.dart';
+import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:illinois/ui/widgets/TabBarWidget.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
@@ -92,7 +92,7 @@ class _GroupPostCreatePanelState extends State<GroupPostCreatePanel>{
                     Flexible(
                       flex: 1,
                       child: RoundedButton(
-                        label: Localization().getStringEx('panel.group.detail.post.create.button.send.title', 'Send'),
+                        label: Localization().getStringEx('panel.group.detail.post.create.button.send.title', 'Send')!,
                         borderColor: Styles().colors!.fillColorSecondary,
                         textColor: Styles().colors!.fillColorPrimary,
                         backgroundColor: Styles().colors!.white,
@@ -101,7 +101,7 @@ class _GroupPostCreatePanelState extends State<GroupPostCreatePanel>{
                     Flexible(
                       flex: 1,
                       child: RoundedButton(
-                        label: Localization().getStringEx('panel.group.detail.post.create.button.cancel.title', 'Cancel'),
+                        label: Localization().getStringEx('panel.group.detail.post.create.button.cancel.title', 'Cancel')!,
                         borderColor: Styles().colors!.textSurface,
                         textColor: Styles().colors!.fillColorPrimary,
                         backgroundColor: Styles().colors!.white,