+++ lib/ui/settings/SettingsLoginNetIdPanel.dart
@@ -23,7 +23,7 @@ import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:illinois/ui/onboarding/OnboardingBackButton.dart';
-import 'package:illinois/ui/widgets/RoundedButton.dart';
+import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 
 class SettingsLoginNetIdPanel extends StatefulWidget{
   
@@ -104,7 +104,7 @@ class _SettingsLoginNetIdPanelState extends State<SettingsLoginNetIdPanel> imple
                   child: Padding(
                     padding: EdgeInsets.all(24),
                     child: RoundedButton(
-                        label: Localization().getStringEx('panel.settings.login.netid.button.continue.title', 'Log in with NetID'),
+                        label: Localization().getStringEx('panel.settings.login.netid.button.continue.title', 'Log in with NetID')!,
                         hint: Localization().getStringEx('panel.settings.login.netid.button.continue.hint', ''),
                         borderColor: Styles().colors!.fillColorSecondary,
                         backgroundColor: Styles().colors!.background,