+++ lib/ui/wallet/WalletPanel.dart
@@ -35,7 +35,7 @@ import 'package:illinois/ui/settings/SettingsIlliniCashPanel.dart';
 import 'package:illinois/ui/settings/SettingsMealPlanPanel.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/VerticalTitleContentSection.dart';
-import 'package:illinois/ui/widgets/RoundedButton.dart';
+import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 
 class WalletPanel extends StatefulWidget{
@@ -196,7 +196,7 @@ class _WalletPanelState extends State<WalletPanel> implements NotificationsListe
 
   Widget _buildLoginNetIdButton() {
     return RoundedButton(
-      label: Localization().getStringEx('panel.wallet.button.connect.netid.title', 'Connect NetID'),
+      label: Localization().getStringEx('panel.wallet.button.connect.netid.title', 'Connect NetID')!,
       hint: Localization().getStringEx('panel.wallet.button.connect.netid.hint', ''),
       backgroundColor: Styles().colors!.surface,
       fontSize: 16.0,
@@ -222,7 +222,7 @@ class _WalletPanelState extends State<WalletPanel> implements NotificationsListe
 
   Widget _buildLoginPhoneOrEmailButton() {
     return RoundedButton(
-        label: Localization().getStringEx('panel.wallet.button.connect.phone_or_email.title', 'Login By Email or Phone'),
+        label: Localization().getStringEx('panel.wallet.button.connect.phone_or_email.title', 'Login By Email or Phone')!,
         hint: Localization().getStringEx('panel.wallet.button.connect.phone_or_email.hint', ''),
         backgroundColor: Styles().colors!.surface,
         fontSize: 16.0,
@@ -437,7 +437,7 @@ class _WalletPanelState extends State<WalletPanel> implements NotificationsListe
             Container(height: 5,),
             Semantics(explicitChildNodes: true,child:
               RoundedButton(
-                label: Localization().getStringEx("panel.wallet.button.use_bus_pass.title", "Use bus pass"),
+                label: Localization().getStringEx("panel.wallet.button.use_bus_pass.title", "Use bus pass")!,
                 hint: Localization().getStringEx("panel.wallet.button.use_bus_pass.hint", ""),
                 textColor: Styles().colors!.fillColorPrimary,
                 backgroundColor: Styles().colors!.white,
@@ -485,7 +485,7 @@ class _WalletPanelState extends State<WalletPanel> implements NotificationsListe
             Container(height: 5,),
             Semantics(explicitChildNodes: true,child:
               RoundedButton(
-                label: Localization().getStringEx("panel.wallet.button.use_id.title", "Use ID"),
+                label: Localization().getStringEx("panel.wallet.button.use_id.title", "Use ID")!,
                 hint: Localization().getStringEx("panel.wallet.button.use_id.hint", ""),
                 textColor: Styles().colors!.fillColorPrimary,
                 backgroundColor: Styles().colors!.white,